Add absolute number (bytes) changed & max staleness for backup status
Add the absolute number of bytes changed for a backup status compared
to the latest successful run, in addition to the percentage of change
for the full view.

Also add the number of max days of age for a particular latest backup
before it is marked as stale, as configured- each backup job
can have a different expected age (2 days for snapshots, 8 days for
metadata, misc and read-write content dumps and 3 years for read-only
content dumps.

Bug: T283017
Change-Id: Ie5033f51774126b368a0ef47fd2f1b3b7fbe7e5d