pampinus: Fix bugs with codfw-only sections & very small backups
There was a bug with very small backups not showing its status
properly- this was fixed.
Also there was no way to handle codfw-only backups, this has been
added.
In any case, the input configuration should be unified between
icinga/puppet and the dashboard, but still working on that
formatting.

Tweak the lenght of the html template lightly.

Bug: T313582
Change-Id: Ieda203f46546087b7aba6c07179a09ad32cc1065